refactor: 무비타임, 리뷰타임의 같은 슬라이더 애니메이션을 useSliderAnimation 훅으로 분리
